requirements.txt
catboost==1.2
pandas==2.0.2
pycodestyle==2.10.0
pymysql==1.0.3
python-dotenv==1.0.0
scikit-learn==1.2.2
